requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
python-magic==0.4.27
pillow>=10.3.0
pydantic==2.5.3
